+++ src/test/java/io/github/intoto/helpers/provenancev01/IntotoHelperTest.java
@@ -1,4 +1,4 @@
-package io.github.intoto.helpers;
+package io.github.intoto.helpers.provenancev01;
 
 import static io.github.intoto.utilities.KeyUtilities.readPrivateKey;
 import static io.github.intoto.utilities.KeyUtilities.readPublicKey;
@@ -13,16 +13,17 @@ import io.github.intoto.dsse.helpers.SimpleECDSASigner;
 import io.github.intoto.dsse.helpers.SimpleECDSAVerifier;
 import io.github.intoto.dsse.models.IntotoEnvelope;
 import io.github.intoto.exceptions.InvalidModelException;
+import io.github.intoto.helpers.IntotoHelper;
 import io.github.intoto.implementations.FakeSigner;
 import io.github.intoto.models.DigestSetAlgorithmType;
 import io.github.intoto.models.Predicate;
 import io.github.intoto.models.Statement;
 import io.github.intoto.models.Subject;
-import io.github.intoto.slsa.models.Builder;
-import io.github.intoto.slsa.models.Material;
-import io.github.intoto.slsa.models.Provenance;
-import io.github.intoto.slsa.models.Recipe;
-import io.github.intoto.utilities.IntotoStubFactory;
+import io.github.intoto.slsa.models.v01.Builder;
+import io.github.intoto.slsa.models.v01.Material;
+import io.github.intoto.slsa.models.v01.Provenance;
+import io.github.intoto.slsa.models.v01.Recipe;
+import io.github.intoto.utilities.provenancev01.IntotoStubFactory;
 import java.io.File;
 import java.nio.charset.StandardCharsets;
 import java.security.InvalidKeyException;
