+++ src/test/java/io/github/intoto/utilities/TestEnvelopeVerify.java
@@ -18,6 +18,8 @@ import java.security.PublicKey;
 import java.security.Security;
 import java.util.Base64;
 import java.util.Objects;
+
+import io.github.intoto.utilities.provenancev01.TestEnvelopeGenerator;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 
 public class TestEnvelopeVerify {
